hospital-platform/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
motor==3.4.0
pymongo==4.7.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
email-validator==2.2.0